+++ src/netlist/netlist_utils.cpp
@@ -7,6 +7,7 @@
 #include "hal_core/netlist/netlist_factory.h"
 #include "hal_core/utilities/log.h"
 
+#include <deque>
 #include <queue>
 #include <unordered_set>
 
@@ -884,6 +885,85 @@ namespace hal
             }
 
             return true;
+        }
+
+        std::vector<Gate*> get_gate_chain(Gate* start_gate, const std::string& pin, const std::function<bool(const Gate*)>& filter)
+        {
+            if (!filter(start_gate))
+            {
+                return {};
+            }
+
+            std::deque<Gate*> gate_chain = {start_gate};
+
+            const GateType* target_type = start_gate->get_type();
+
+            // move forward
+            bool found_next_gate;
+            const Gate* current_gate = start_gate;
+            do
+            {
+                found_next_gate = false;
+
+                // check all successors of current gate
+                std::vector<Endpoint*> successors = current_gate->get_successors();
+                for (const Endpoint* suc_ep : successors)
+                {
+                    // check pin name and gate type
+                    Gate* suc_gate = suc_ep->get_gate();
+                    if (suc_ep->get_pin() == pin && suc_gate->get_type() == target_type)
+                    {
+                        // check filter condition
+                        if (!filter(suc_gate))
+                        {
+                            continue;
+                        }
+
+                        // check if more than one successor
+                        if (!found_next_gate)
+                        {
+                            gate_chain.push_back(suc_gate);
+                            current_gate    = suc_gate;
+                            found_next_gate = true;
+                            log_debug("netlist_utils", "found successor gate with ID {}.", suc_gate->get_id());
+                        }
+                        else
+                        {
+                            log_error("netlist_utils",
+                                      "detected more than one valid successor gate for gate '{}' with ID {} in netlist with ID {}.",
+                                      suc_gate->get_name(),
+                                      suc_gate->get_id(),
+                                      suc_gate->get_netlist()->get_id());
+                            return {start_gate};
+                        }
+                    }
+                }
+            } while (found_next_gate);
+
+            // move backwards
+            current_gate = start_gate;
+            do
+            {
+                found_next_gate = false;
+                if (Endpoint* pred_ep = current_gate->get_predecessor(pin); pred_ep != nullptr)
+                {
+                    Gate* pred_gate = pred_ep->get_gate();
+
+                    if (pred_gate->get_type() == target_type)
+                    {
+                        // check filter condition
+                        if (filter(pred_gate))
+                        {
+                            gate_chain.push_front(pred_gate);
+                            current_gate    = pred_gate;
+                            found_next_gate = true;
+                            log_debug("netlist_utils", "found predecessor gate with ID {}.", pred_gate->get_id());
+                        }
+                    }
+                }
+            } while (found_next_gate);
+
+            return std::vector<Gate*>(gate_chain.begin(), gate_chain.end());
         }
     }    // namespace netlist_utils
 }    // namespace hal